Remove value type from auxiliary functions
Since we are dealing with symbolics, the values can be evaluated later.

+++ src/functions.jl
@@ -1,7 +1,7 @@
 using IfElse
 
 interpolate(x, x₁, xₙ, y₁, yₙ) = y₁ + (x - x₁) * ((yₙ - y₁) / (xₙ - x₁))
-function interpolate(x::Float64, yvalues::Tuple{Vararg{Float64}}, xrange::Tuple{Float64, Float64})::Float64
+function interpolate(x, yvalues::Tuple{Vararg{Float64}}, xrange::Tuple{Float64, Float64})
     xvalues = LinRange(xrange[1], xrange[2], length(yvalues))
 
     # y gets the min y value if less than the min x range
@@ -16,6 +16,6 @@ function interpolate(x::Float64, yvalues::Tuple{Vararg{Float64}}, xrange::Tuple{
     return y
 end
 
-clip(returnifgte::Float64, returniflt::Float64, inputvalue::Float64, threshold::Float64) = IfElse.ifelse(inputvalue ≥ threshold, returnifgte, returniflt)
-step(inputvalue::Float64, returnifgte::Float64, threshold::Float64) = clip(returnifgte, zero(returnifgte), inputvalue, threshold)
-switch(returnifzero::Float64, returnifnotzero::Float64, inputvalue::Float64) = IfElse.ifelse(isapprox(inputvalue, zero(inputvalue); atol=1e-16), returnifzero, returnifnotzero)
+clip(returnifgte, returniflt, inputvalue, threshold) = IfElse.ifelse(inputvalue ≥ threshold, returnifgte, returniflt)
+step(inputvalue, returnifgte, threshold) = clip(returnifgte, zero(returnifgte), inputvalue, threshold)
+switch(returnifzero, returnifnotzero, inputvalue) = IfElse.ifelse(isapprox(inputvalue, zero(inputvalue); atol=1e-16), returnifzero, returnifnotzero)
